회원가입, 로그인

+++ src/main/java/com/example/velogclonebe/service/ArticleService.java
@@ -2,6 +2,8 @@ package com.example.velogclonebe.service;
 
 
 import com.example.velogclonebe.domain.dto.request.ArticleRequestDto;
+import com.example.velogclonebe.domain.dto.request.ArticleUpdateRequestDto;
+import com.example.velogclonebe.domain.dto.response.ArticleListResponseDto;
 import com.example.velogclonebe.domain.dto.response.ArticleResponseDto;
 import com.example.velogclonebe.domain.dto.response.CommentGetResponseDto;
 import com.example.velogclonebe.domain.entity.Article;
@@ -23,33 +25,54 @@ public class ArticleService {
     private final ArticleRepository articleRepository;
     private final CommentRepository commentRepository;
 
+
     // 게시글 리스트 조회
     @Transactional
-    public List<Article> getArticles() {
+    public List<ArticleListResponseDto> getArticles() {
+
         List<Article> articleList = articleRepository.findAllByOrderByCreatedAtDesc();
-        return articleList;
+        // List<Article> articleList = articleRepository.findAllAndCountComment();
+
+        List<ArticleListResponseDto> responseDto = articleList.stream()
+                .map(article -> new ArticleListResponseDto(article, commentRepository.findAllByArticleOrderByCreatedAtDesc(article)))
+                .collect(Collectors.toList());
+
+        return responseDto;
+
     }
 
     // 게시글 작성
     @Transactional
-    public void setArticle(ArticleRequestDto articleRequestDto) {
-        Article article = new Article(articleRequestDto);
+    public void setArticle(ArticleRequestDto articleRequestDto, String username) {
+        Article article = new Article(articleRequestDto, username);
         articleRepository.save(article);
     }
 
     // 게시글 수정
     @Transactional
-    public void updateArticle(Long articleId, ArticleRequestDto articleRequestDto) {
+    public void updateArticle(Long articleId, ArticleUpdateRequestDto articleUpdateRequestDto, String username) {
+
         Article article = articleRepository.findById(articleId).orElseThrow(
-                ()-> new ApiRequestException("해당 글이 존재하지 않습니다."));
-        article.update(articleRequestDto);
+                ()  -> new ApiRequestException("해당 글이 존재하지 않습니다."));
+
+        if (!article.getUsername().equals(username)) {
+            throw new ApiRequestException("수정 권한이 없습니다.");
+        }
+
+        article.update(articleUpdateRequestDto);
     }
 
     // 게시글 삭제
     @Transactional
-    public void deleteArticle(Long articleId, ArticleRequestDto articleRequestDto) {
+    public void deleteArticle(Long articleId, String username) {
+
         Article article = articleRepository.findById(articleId).orElseThrow(
                 ()->new ApiRequestException("해당 글이 존재하지 않습니다."));
+
+        if (!article.getUsername().equals(username)) {
+            throw new ApiRequestException("삭제 권한이 없습니다.");
+        }
+
 
         // 글 삭제시 해당 글에 연관된 댓글들 삭제 (commentRepository 미반영)
         commentRepository.deleteAllByArticle(article);
@@ -57,18 +80,30 @@ public class ArticleService {
     }
 
     // 게시글 상세페이지 요청 처리
+    @Transactional
     public ArticleResponseDto getArticleDetail(Long articleId) {
+
         Article article = articleRepository.findById(articleId).orElseThrow(
                 ()->new ApiRequestException("해당 글이 존재하지 않습니다."));
+
         // 게시글에 해당하는 댓글 리스트로 받아옴
         List<Comment> commentList = commentRepository.findAllByArticleOrderByCreatedAtDesc(article);
+
         // Dto로 넘겨주기 위한 작업
         List<CommentGetResponseDto> commentDtoList = commentList.stream()
                 .map(CommentGetResponseDto::new)
                 .collect(Collectors.toList());
+
         // 게시글 응답을 위해 articleresponsedto를 생성해서 반환
         ArticleResponseDto articleResponseDto = new ArticleResponseDto(article, commentDtoList);
         return articleResponseDto;
     }
 
+    // 검색
+    @Transactional
+    public List<Article> getSearchArticles(String keyword) {
+        List<Article> searchedArticles = articleRepository.findByTitleContaining(keyword);
+
+        return searchedArticles;
+    }
 }
